chore: split model declarations
